requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
pypdf
pdfplumber
langchain
langchain-community
langchain-openai
chromadb
tiktoken
python-dotenv
requests
